ConvolutionalQuantumNetwork/requirements.txt
numpy
qiskit
torch
torchvision
tqdm
tensorflow
matplotlibx